fixed:migration failed error

+++ src/db/migrations/1765541196119-add-leather-color-to-order-items.ts
@@ -4,10 +4,21 @@ export class AddLeatherColorToOrderItems1765541196119 implements MigrationInterf
     name = 'AddLeatherColorToOrderItems1765541196119'
 
     public async up(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.query(`ALTER TABLE "order_items" ADD "leather_color" character varying`);
+        const table = await queryRunner.getTable("order_items");
+        const column = table?.findColumnByName("leather_color");
+        
+        if (!column) {
+            await queryRunner.query(`ALTER TABLE "order_items" ADD "leather_color" character varying`);
+        }
     }
 
     public async down(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.query(`ALTER TABLE "order_items" DROP COLUMN "leather_color"`);
+        const table = await queryRunner.getTable("order_items");
+        const column = table?.findColumnByName("leather_color");
+        
+        if (column) {
+            await queryRunner.query(`ALTER TABLE "order_items" DROP COLUMN "leather_color"`);
+        }
     }
 }
+
